% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDF.R
\name{inspectDF}
\alias{inspectDF}
\title{obtain rules as data.frame from arules instead of original inspect().}
\usage{
inspectDF(rules, sep = ",")
}
\arguments{
\item{rules}{a rule object of  \code{\link{arules}}.}

\item{sep}{string. items in rule string will be separated by sep.}
}
\value{
a data frame of rules.
}
\description{
obtain rules as data.frame from arules instead of original inspect().
}
\examples{
\dontrun{
data(Groceries)

pars <- list(confidence=0.01, support=0.01, maxlen=5, minlen=2)
glo.apriori <- apriori(Groceries, parameter = pars)

glo.inspectDF  <- inspectDF(glo.apriori)
glo.inspectDF \%>\% head
}

}
