% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotEdgeWithRule}
\alias{plotEdgeWithRule}
\title{plot edge.data.frame with rules as graph.}
\usage{
plotEdgeWithRule(edges, rules, adujust.support.size = 200, label = "",
  cex = 1, ...)
}
\arguments{
\item{edges}{a data.frame to describe edges and nodes, for inputs to graph.data.frame.}

\item{rules}{a data.frame to describe weights for edge.}

\item{adujust.support.size}{adujust.support.size}

\item{label}{labels for item node.}

\item{cex}{label size for ittem node.}

\item{...}{arguments path to plot.igraph()}
}
\description{
plot edge.data.frame with rules as graph.
}
