% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotRuleGraph}
\alias{plotRuleGraph}
\title{plot rules.data.frame as graph.}
\usage{
plotRuleGraph(rules.data.frame, adujust.support.size = 50, label = "",
  cex = 1, ...)
}
\arguments{
\item{rules.data.frame}{rules.data.frame.}

\item{adujust.support.size}{adujust.support.size}

\item{label}{labels for item node.}

\item{cex}{label size for ittem node.}

\item{...}{arguments path to plot.igraph()}
}
\description{
plot rules.data.frame as graph.
}
