% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report summary statistics for a data frame}
\usage{
report(df, top_n = NULL)
}
\arguments{
\item{df}{A data frame}

\item{top_n}{The number of rows to print for summaries \code{report_na}, \code{report_cor}, \code{report_imbalance}, \code{report_association}, Default \code{NULL} prints everything.}
}
\value{
Prints statistics to the console.
}
\description{
Report summary statistics for a data frame
}
\details{
\code{report} is a wrapper for the all of the individual reporting functions \code{report_space}, \code{report_types}, \code{report_na}, \code{report_cor}, \code{report_imbalance}, \code{report_association}.
}
\examples{
report(starwars)
}
