% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_space.R
\name{report_mem}
\alias{report_mem}
\title{Report the memory usage of a data frame or compare usage in two data frames.}
\usage{
report_mem(df1, df2 = NULL, top = NULL, show_plot = FALSE)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{An optional second data frame for comparison.}

\item{top}{Positive integer specifying the number of rows to print in the 
output tibble.  Useful when the input data frame has many columns.  
\code{top = NULL}, the default causes everything to be returned.}

\item{show_plot}{Logical argument determining whether plot is generated 
in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
A tibble summarising the memory usage of each column
 in one or two data frames.
}
\description{
Report the memory usage of a data frame or compare usage in two data frames.
}
\details{
When \code{df2 = NULL}, a tibble is returned with the columns: \code{col_name}
the columns in \code{df1}, \code{size} and \code{pcnt} contain the memory usage 
of each column in \code{df1}.  The tibble is sorted in descending order of \code{size}.

When a second data frame \code{df2} is specified, column sizes are 
tabulated for both data frames to enable comparison.  
A full join is performed between size summary table for the two data frames: where
a column exists in one but not the other, \code{size} and \code{pcnt} some 
cells return \code{NA}.
}
\examples{
data("starwars", package = "dplyr")
# get tibble of column memory usage for the starwars data
report_mem(starwars)
# get column memory usage and show as barplot
report_mem(starwars, show_plot = TRUE)
# compare memory usage 
report_space(starwars, starwars[1:10, -3])
}
