% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_na.R
\name{report_na}
\alias{report_na}
\title{Report the proportion of each column containing missing values}
\usage{
report_na(df, top_n = NULL, type = "df")
}
\arguments{
\item{df}{A data frame}

\item{top_n}{The number of rows to print for summaries. Default \code{top_n = NULL} prints everything.}

\item{type}{Character specificying report output type.  Default \code{type = "df"} causes report to be returned as a tibble.   \code{type = "console"} causes report to be returned directly to the console.}
}
\value{
Return a \code{tibble} containing the columns \code{col_name}, \code{count_na} and \code{percent_na}.
}
\description{
Report the proportion of each column containing missing values
}
\examples{
report_na(mtcars)
}
