% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_num.R
\name{report_num}
\alias{report_num}
\title{Report and compare the numeric variables within one or two data frames}
\usage{
report_num(df1, df2 = NULL, top = NULL, show_plot = F,
  breakseq = NULL, breaks = 20)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for comparing categorical levels.
Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. 
Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition 
to tibble output.  Default is \code{FALSE}.}

\item{breakseq}{For internal use only.  Argument that accepts a pre-specified set of 
break points, default is \code{NULL}.}

\item{breaks}{Optional argument determining how breaks are constructed for 
histograms to use when comparing numeric data frame features.  
This takes the same values as \code{hist(..., breaks)}.  See \code{?hist} 
for more details.}
}
\value{
A \code{tibble} containing statistical summaries of the numeric 
columns of \code{df1}, or comparing the histograms of \code{df1} and \code{df2}.
}
\description{
Report and compare the numeric variables within one or two data frames
}
\details{
If only \code{df1} is specified, then the tibble returned will have the following columns:
\itemize{
  \item \code{col_name} the columns contained in \code{df1}
  \item \code{min}, \code{q1}, \code{median}, \code{mean}, \code{q3}, \code{max} and \code{sd}: 
 the minimum, lower quartile, median, mean, upper quartile, maximum and standard deviation 
 for each numeric column.
  \item \code{pcnt_na} the percentage of each numeric feature that is missing
  \item \code{hist} a list of tibbles containing the relative frequency of values in a 
  set of discrete bins for each column.
}
}
\examples{
data("starwars", package = "dplyr")
# show summary statistics for starwars
report_num(starwars)
# show a visualisation too - limit number of bins
report_num(starwars, breaks = 10)
# compare two data frames
report_num(starwars, starwars[-c(1:10), ], breaks = 10, show_plot = TRUE)
}
