% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_numeric.R
\name{report_numeric}
\alias{report_numeric}
\title{Report and compare the numeric variables within one or two data frames}
\usage{
report_numeric(df1, df2 = NULL, top = NULL, show_plot = F,
  breaks = NULL)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for comparing categorical levels  Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.  Default is \code{FALSE}.}

\item{breaks}{Optional argument specifying the breaks to use when comparing numeric data frame features.}
}
\value{
If \code{df2 = NULL} then is a \code{tibble} containing the names of numeric columns (\code{col_name}).
}
\description{
Report and compare the numeric variables within one or two data frames
}
\examples{
data("starwars", package = "dplyr")
report_numeric(starwars)
}
