% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_space.R
\name{report_space}
\alias{report_space}
\title{Report the memory usage of a data frame}
\usage{
report_space(df1, df2 = NULL, top = NULL, show_plot = FALSE)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for comparing memory usage with.  Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
Prints the proportion of overall memory used by each column and the total usage.
}
\description{
Report the memory usage of a data frame
}
\examples{
# get tibble of column memory usage for the starwars data
report_space(starwars)
# get column memory usage and show as barplot
report_space(starwars, show_plot = TRUE)
}
