% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_types.R
\name{report_types}
\alias{report_types}
\title{Report column types of a data frame}
\usage{
report_types(df, type = "df")
}
\arguments{
\item{df}{A data frame}

\item{type}{Character specificying report output type.  Default \code{type = "df"} causes report to be returned as a tibble.   \code{type = "console"} causes report to be returned directly to the console.}
}
\value{
Prints the proportion of columns with each type.
}
\description{
Report column types of a data frame
}
\examples{
report_types(starwars)
}
