% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEdges.R
\name{toEdges}
\alias{toEdges}
\title{convert rule data.frame to input for graph.data.frame (with from and to colmn).}
\usage{
toEdges(rules.data.frame, sep = ",")
}
\arguments{
\item{rules.data.frame}{an object of inspectDf, or data.frame with LHS & RHS colmn.}

\item{sep}{string to separate items from rule string.}
}
\value{
a data frame with from and to colmn for graph.data.frame inputs.
}
\description{
convert rule data.frame to input for graph.data.frame (with from and to colmn).
}
\examples{
\dontrun{
data(Groceries)

pars <- list(confidence=0.01, support=0.01, maxlen=5, minlen=2)
glo.apriori <- apriori(Groceries, parameter = pars)

glo.inspectDF  <- inspectDF(glo.apriori)
glo.inspectDF \%>\% head

#' glo.inspectDF \%>\% toEdges \%>\% head
}

}
