% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_association.R
\name{report_association}
\alias{report_association}
\title{Report Goodman and Kruskal's tau for association between categorical features}
\usage{
report_association(df1, df2 = NULL, top = NULL, show_plot = F)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{An optional second data frame for comparing with \code{df2}.  Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
Return a \code{tibble} containing the columns \code{col_1}, \code{col_2}, \code{pair} and \code{k_tau}.
}
\description{
Report Goodman and Kruskal's tau for association between categorical features
}
\examples{
report_association(starwars)
}
