% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_levels.R
\name{report_cat}
\alias{report_cat}
\title{Report and compare the levels within each categorical feature}
\usage{
report_cat(df1, df2 = NULL, top = NULL, show_plot = FALSE)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for comparing categorical levels  Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
If \code{df2 = NULL} then is a \code{tibble} containing the names of categorical columns (\code{col_name}), the number of levels within each (\code{n_levels}), the most common level (\code{dom_level}), the percentage occurrence of the most common feature (\code{dom_percent}) and a list of tibbles containing the percentage appearance of each feature (\code{levels}).
}
\description{
Report and compare the levels within each categorical feature
}
\examples{
data("starwars", package = "dplyr")
report_cat(starwars)
}
