% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cor.R
\name{report_cor}
\alias{report_cor}
\title{Report the Pearson's correlation coefficient for each pair of numeric columns}
\usage{
report_cor(df, df2 = NULL, top_n = NULL, type = "df")
}
\arguments{
\item{df}{A data frame containing numeric columns}

\item{df2}{An optional second data frame for comparing correlation coefficients with.  Defaults to \code{NULL}.}

\item{top_n}{The number of rows to print for summaries. Default \code{top_n = NULL} prints everything.}

\item{type}{Character specificying report output type.  Default \code{type = "df"} causes report to be returned as a tibble.   \code{type = "console"} causes report to be returned directly to the console.}
}
\value{
Return a \code{tibble} containing the columns \code{col_1}, \code{col_2} and \code{pair} and \code{correlation}.  The report contains only the upper triangle of the correlation matrix.  The tibble is sorted by descending absolute value in the \code{correlation} column.
}
\description{
Report the Pearson's correlation coefficient for each pair of numeric columns
}
\details{
When the second data frame \code{df2} is specified, correlations are tabulated for both data frames, and where a pair of numeric columns with the same names appear in both, a p-value is provided which test tests whether their correlations coefficients are equal.
}
\examples{
report_cor(starwars)
report_cor(starwars, starwars[1:10, ])
}
