% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cor.R
\name{report_cor}
\alias{report_cor}
\title{Report the Pearson's correlation coefficient for each pair of numeric columns}
\usage{
report_cor(df, top_n = NULL, type = "df")
}
\arguments{
\item{df}{A data frame}

\item{top_n}{The number of rows to print for summaries. Default \code{top_n = NULL} prints everything.}

\item{type}{Character specificying report output type.  Default \code{type = "df"} causes report to be returned as a tibble.   \code{type = "console"} causes report to be returned directly to the console.}
}
\value{
Return a \code{tibble} containing the columns \code{col_1}, \code{col_2} and \code{pair} and \code{correlation}.  The report contains only the upper triangle of the correlation matrix.  The tibble is sorted by descending absolute value in the \code{correlation} column.
}
\description{
Report the Pearson's correlation coefficient for each pair of numeric columns
}
\examples{
report_cor(mtcars)
}
