% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mem.R
\name{report_mem}
\alias{report_mem}
\title{Report and compare the memory usage in one or two dataframes.}
\usage{
report_mem(df1, df2 = NULL, show_plot = FALSE)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{An optional second data frame for comparing column sizes.
Defaults to \code{NULL}.}

\item{show_plot}{Logical argument determining whether plot is returned
in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
A tibble summarising and comparing the columnwise memory usage 
for one or a pair of data frames.
}
\description{
Report and compare the memory usage in one or two dataframes.
}
\details{
When a single data frame is specified, a tibble is returned which 
contains columnwise memory usage in descending order of size:
\itemize{
  \item \code{col_name} character vector containing column names of \code{df1}.
  \item \code{size} character vector containing memory usage of each column.
  \item \code{pcnt} the percentage of total memory usage used by each column.
}
When both \code{df1} and \code{df2} are specified, column sizes are jointly 
tabulated for both data frames, by performing a full join by \code{col_name}.  
Rows are sorted in descending order of size as they appear in \code{df1}:
\itemize{
  \item \code{col_name} character vector containing column names of \code{df1}
  and \code{df2}.
  \item \code{size_} character vector containing memory usage of each column in
  each of \code{df1} and \code{df2}.
  \item \code{pcnt_} the percentage of total memory usage of each column within 
  each of \code{df1} and \code{df2}.
}
}
\examples{
data("starwars", package = "dplyr")
# get tibble of column memory usage for the starwars data
report_mem(starwars)
# get column memory usage and show as barplot
report_mem(starwars, show_plot = TRUE)
# compare memory usage 
report_mem(starwars, starwars[1:10, -3])
}
