% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_na.R
\name{report_na}
\alias{report_na}
\title{Report the rate of missingness in each column of a dataframe}
\usage{
report_na(df1, df2 = NULL, top = NULL, show_plot = FALSE,
  alpha = 0.05)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for making comparisons of missingness across columns.
Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} 
prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.
Default is \code{FALSE}.}

\item{alpha}{Alpha level for performing significance tests.  Defaults to 0.05.}
}
\value{
Return a \code{tibble} summarising columnwise missingness by presence of 
\code{NA}.  Output contains the columns \code{col_name}, \code{cnt_na} and \code{pcnt}.
}
\description{
Report the rate of missingness in each column of a dataframe
}
\details{
When a second data frame \code{df2} is specified, the tibble returned compares 
missing for both dataframes.  The \code{p_value} results from a a chi-square test 
of the null hypothesis that the rate of missingness is the same in in \code{df1} 
and \code{df2}.
}
\examples{
data("starwars", package = "dplyr")
# missingness in starwars data
report_na(starwars)
report_na(starwars, show_plot = TRUE)
# compare missingness 
report_na(starwars, starwars[1:30, ])
}
