% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_na.R
\name{report_na}
\alias{report_na}
\title{Report the proportion of each column containing missing values}
\usage{
report_na(df1, df2 = NULL, top = NULL, show_plot = F)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{An optional second data frame for comparing missing values with.  Defaults to \code{NULL}.}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}

\item{show_plot}{Logical determining whether to show a plot in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
Return a \code{tibble} containing the columns \code{col_name}, \code{count_na} and \code{percent_na}.
}
\description{
Report the proportion of each column containing missing values
}
\details{
When the second data frame \code{df2} is specified, the missingness is tabulated for both data frames, and where a pair of columns are common to both data frames a p-value is calculated for the equivalence of the proportion of missing values.
}
\examples{
report_na(starwars)
}
