% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_space.R
\name{report_space}
\alias{report_space}
\title{Report the memory usage of a data frame}
\usage{
report_space(df1, df2 = NULL, top = NULL)
}
\arguments{
\item{df1}{A data frame}

\item{top}{The number of rows to print for summaries. Default \code{top = NULL} prints everything.}
}
\value{
Prints the proportion of overall memory used by each column and the total usage.
}
\description{
Report the memory usage of a data frame
}
\examples{
report_space(starwars)

}
