% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_types.R
\name{report_types}
\alias{report_types}
\title{Report column types in a data frame or compare types in two data frames.}
\usage{
report_types(df1, df2 = NULL, show_plot = FALSE)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{An optional second data frame for comparison.}

\item{show_plot}{Logical argument determining whether plot is generated 
in addition to tibble output.  Default is \code{FALSE}.}
}
\value{
A tibble summarising the count and percentage of different 
column types for one or a pair of data frames.
}
\description{
Report column types in a data frame or compare types in two data frames.
}
\details{
When \code{df2 = NULL}, a tibble is returned with the columns
\itemize{
  \item \code{type} the column types contained in \code{df1}
  \item \code{cnt} the number of columns with each type
  \item \code{pcnt} the percentage of columns with each type
  \item \code{col_name} the names of columns with a specific type
}
When a second data frame \code{df2} is specified, column type summaries 
are tabulated for both data frames to enable comparison of contents. 
The resulting tibble has the columns
\itemize{
  \item \code{type} column types present in either data frame
  \item \code{cnt_...} pair of columns containing count of columns with 
  each type - the data frame name are appended.
  \item \code{pcnt_...} pair of columns containing the percentage of 
  columns with each type - the data frame name are appended.
}
}
\examples{
data("starwars", package = "dplyr")
# get tibble of column types for the starwars data
report_types(starwars)
# get column types and show as barplot
report_types(starwars, show_plot = TRUE)
}
