% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech-data.R
\docType{data}
\name{tech}
\alias{tech}
\title{Tech stocks closing prices}
\format{
A \code{dataframe} with 3158 rows and 6 columns.
}
\source{
Data gathered using the \href{https://github.com/joshuaulrich/quantmod}{quantmod}
package.
}
\usage{
data(tech)
}
\description{
Daily closing stock prices of the three tech companies Microsoft, 
Apple and IBM between 2007 and 2019.
}
\examples{
data(tech)
head(tech)
# NOT RUN - change in correlation over time
# library(dplyr)
# tech_grp <- tech \%>\% 
#         group_by(year) \%>\%
#         inspect_cor()
# tech_grp \%>\% show_plot()    
}
\keyword{datasets}
